requirement.txt
django
djangorestframework
bcrypt
json
sys
os
